refactor: split main.rs into cli and resolve modules
Move the clap derive types (Cli, Commands, GenerateCommands) into cli.rs and
the compose-file/base-dir/project-name resolution plus its tests into
resolve.rs, bringing main.rs under the 500-line limit. Behavior unchanged.
