+++ internal/main.rs
@@ -1,255 +1,16 @@
 //! `podup` — docker-compose to Podman translator CLI.
 
-use std::path::{Path, PathBuf};
+use std::path::Path;
 use std::process;
 
-use clap::{Parser, Subcommand};
+use clap::Parser;
 use tracing_subscriber::EnvFilter;
 
-#[derive(Parser)]
-#[command(
-	name = "podup",
-	version,
-	about = "docker-compose translator for Podman"
-)]
-struct Cli {
-	/// Path to the compose file. May also be set via `COMPOSE_FILE`. When
-	/// unset, the compose-spec precedence list is probed in the current
-	/// directory (compose.yaml, compose.yml, docker-compose.yaml,
-	/// docker-compose.yml).
-	#[arg(short, long)]
-	file: Vec<PathBuf>,
+mod cli;
+mod resolve;
 
-	/// Project name (used as a prefix for container names). May also be set via
-	/// `COMPOSE_PROJECT_NAME`. When unset, the compose-spec precedence applies:
-	/// the top-level `name:` field, then the sanitized basename of the project
-	/// directory.
-	#[arg(short, long, env = "COMPOSE_PROJECT_NAME")]
-	project: Option<String>,
-
-	/// Podman socket path (overrides auto-detection and PODMAN_SOCKET env).
-	#[arg(long, env = "PODMAN_SOCKET")]
-	socket: Option<String>,
-
-	/// Active profiles (comma-separated).  May also be set via `COMPOSE_PROFILES`.
-	#[arg(long, value_delimiter = ',', global = true)]
-	profile: Vec<String>,
-
-	/// Base directory for resolving relative paths (env_file, build context,
-	/// bind mounts, config/secret file sources). Defaults to the directory
-	/// containing the compose file.
-	#[arg(long, global = true)]
-	project_directory: Option<PathBuf>,
-
-	/// Additional env file(s) loaded into the variable map used for
-	/// interpolation. May be given multiple times; later files win. The
-	/// process environment and a project `.env` still take precedence.
-	#[arg(long = "env-file", global = true)]
-	env_file: Vec<String>,
-
-	#[command(subcommand)]
-	command: Commands,
-}
-
-#[derive(Subcommand)]
-enum Commands {
-	/// Create and start all services.
-	Up {
-		/// Run containers in the background.
-		#[arg(short, long)]
-		detach: bool,
-		/// Build images before starting containers.
-		#[arg(long)]
-		build: bool,
-		/// Watch for file changes and sync/rebuild/restart per develop.watch rules.
-		#[arg(short, long)]
-		watch: bool,
-		/// Remove containers for services not defined in the compose file.
-		#[arg(long)]
-		remove_orphans: bool,
-		/// Do not recreate containers that are already running.
-		#[arg(long)]
-		no_recreate: bool,
-		/// Recreate containers even if their configuration is unchanged.
-		#[arg(long)]
-		force_recreate: bool,
-		/// Do not start linked services (depends_on) of the named services.
-		#[arg(long)]
-		no_deps: bool,
-		/// Bring up only these services (and their transitive depends_on).
-		/// If omitted, brings up every service in the compose file.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Stop and remove containers.
-	Down {
-		/// Also remove named volumes declared in the compose file.
-		#[arg(short = 'v', long)]
-		volumes: bool,
-	},
-	/// Start existing stopped containers.
-	Start {
-		/// Start only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Stop running containers without removing them.
-	Stop {
-		/// Stop only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Build or rebuild service images.
-	Build {
-		/// Build only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Remove stopped service containers.
-	Rm {
-		/// Remove even running containers (stop first).
-		#[arg(short, long)]
-		force: bool,
-		/// Remove only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Send a signal to service containers.
-	Kill {
-		/// Signal to send (default: SIGKILL).
-		#[arg(short, long, default_value = "SIGKILL")]
-		signal: String,
-		/// Signal only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Pause running service containers.
-	Pause {
-		/// Pause only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Resume paused service containers.
-	Unpause {
-		/// Unpause only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Run a one-off command in a new service container.
-	Run {
-		/// Service to run the command against.
-		service: String,
-		/// Remove the container after it exits (default: true).
-		#[arg(long, default_value_t = true)]
-		rm: bool,
-		/// Run container in the background.
-		#[arg(short, long)]
-		detach: bool,
-		/// Set environment variables (KEY=VAL).
-		#[arg(short, long = "env")]
-		env_overrides: Vec<String>,
-		/// Override the container name.
-		#[arg(long)]
-		name: Option<String>,
-		/// Publish the service's declared ports (off by default).
-		#[arg(long)]
-		service_ports: bool,
-		/// Command (and arguments) to run.
-		#[arg(trailing_var_arg = true, allow_hyphen_values = true)]
-		cmd: Vec<String>,
-	},
-	/// Copy files between a service container and the local filesystem.
-	///
-	/// Use SERVICE:PATH for the container side (e.g. `web:/app/data ./local`).
-	Cp {
-		/// Source path. Use SERVICE:PATH for a container path.
-		src: String,
-		/// Destination path. Use SERVICE:PATH for a container path.
-		dst: String,
-	},
-	/// List containers.
-	Ps,
-	/// Display the running processes of service containers.
-	Top {
-		/// Show only these services.
-		#[arg(trailing_var_arg = true)]
-		services: Vec<String>,
-	},
-	/// Print the public port for a port binding of a service container.
-	Port {
-		/// Service name.
-		service: String,
-		/// Private port number.
-		private_port: u16,
-		/// Protocol (tcp or udp).
-		#[arg(long, default_value = "tcp")]
-		proto: String,
-	},
-	/// List images used by services.
-	Images,
-	/// View output from containers.
-	Logs {
-		/// Only show logs for this service.
-		service: Option<String>,
-		/// Follow log output.
-		#[arg(short, long)]
-		follow: bool,
-	},
-	/// Execute a command in a running service container.
-	Exec {
-		/// Service name.
-		service: String,
-		/// Command (and arguments) to execute.
-		#[arg(trailing_var_arg = true, allow_hyphen_values = true)]
-		cmd: Vec<String>,
-	},
-	/// Pull images for all services.
-	Pull,
-	/// Restart services.
-	Restart {
-		/// Only restart this service.
-		service: Option<String>,
-	},
-	/// Print the resolved compose file (after substitution / extends / include).
-	Config,
-	/// Generate declarative artifacts from the compose file.
-	Generate {
-		#[command(subcommand)]
-		kind: GenerateCommands,
-	},
-	/// Watch for file changes and sync/rebuild/restart as configured by develop.watch.
-	Watch,
-	/// Update podup to the latest signed release.
-	///
-	/// Downloads the release binary for this platform and replaces the running
-	/// executable, but only after verifying the release's Ed25519 signature
-	/// against the public key embedded in this build and matching its SHA-256
-	/// checksum. Verification fails closed: a missing key, bad signature, or
-	/// checksum mismatch aborts without touching the installed binary.
-	Update {
-		/// Report whether a newer release exists without installing it.
-		#[arg(long)]
-		check: bool,
-		/// Reinstall even if the latest release is not newer than this build.
-		#[arg(long)]
-		force: bool,
-	},
-}
-
-#[derive(Subcommand)]
-enum GenerateCommands {
-	/// Translate the compose file into Podman Quadlet unit files.
-	///
-	/// Emits one `.container` per service plus `.network` and `.volume` units.
-	/// Without --output the units are printed to stdout; warnings about fields
-	/// with no Quadlet mapping go to stderr.
-	Quadlet {
-		/// Directory to write the unit files into (e.g.
-		/// ~/.config/containers/systemd). Prints to stdout when omitted.
-		#[arg(short, long)]
-		output: Option<PathBuf>,
-	},
-}
+use cli::*;
+use resolve::*;
 
 /// Whether a command creates, destroys, or changes the state of containers and
 /// so must hold the exclusive project lock.
@@ -312,96 +73,6 @@ fn write_quadlet(
 		}
 	}
 	Ok(())
-}
-
-/// Compose-spec file-name precedence, highest first.
-const COMPOSE_FILE_CANDIDATES: [&str; 4] = [
-	"compose.yaml",
-	"compose.yml",
-	"docker-compose.yaml",
-	"docker-compose.yml",
-];
-
-/// Resolve which compose file(s) to load. Explicit `--file` flags win; then the
-/// `COMPOSE_FILE` environment variable (a path-separator-delimited list);
-/// otherwise probe the compose-spec precedence list in the current directory,
-/// falling back to `docker-compose.yml` so a missing-file error names a
-/// sensible path. Multiple files are merged in order, later overriding earlier.
-fn resolve_compose_files(explicit: &[PathBuf]) -> Vec<PathBuf> {
-	if !explicit.is_empty() {
-		return explicit.to_vec();
-	}
-	if let Ok(env) = std::env::var("COMPOSE_FILE") {
-		if !env.is_empty() {
-			let sep = if cfg!(windows) { ';' } else { ':' };
-			return env.split(sep).map(PathBuf::from).collect();
-		}
-	}
-	for candidate in COMPOSE_FILE_CANDIDATES {
-		if Path::new(candidate).is_file() {
-			return vec![PathBuf::from(candidate)];
-		}
-	}
-	vec![PathBuf::from("docker-compose.yml")]
-}
-
-/// Resolve the base directory for relative-path resolution. An explicit
-/// `--project-directory` wins; otherwise it is the directory containing the
-/// compose file (compose-spec default), or the current directory when the
-/// compose file has no parent component.
-fn resolve_base_dir(project_directory: Option<&Path>, file: &Path) -> PathBuf {
-	project_directory
-		.map(Path::to_path_buf)
-		.unwrap_or_else(|| file.parent().map(Path::to_path_buf).unwrap_or_default())
-}
-
-/// Resolve the project name following the compose-spec precedence: an explicit
-/// `-p` / `COMPOSE_PROJECT_NAME` value, then the top-level `name:` field, then
-/// the sanitized basename of the project directory. Explicit values are taken
-/// verbatim; only the directory basename is sanitized.
-fn resolve_project_name(
-	explicit: Option<String>,
-	compose_name: Option<&str>,
-	base_dir: &Path,
-) -> String {
-	if let Some(name) = explicit {
-		return name;
-	}
-	if let Some(name) = compose_name {
-		return name.to_string();
-	}
-	// An empty base dir means a bare compose filename in the current directory;
-	// canonicalize `.` so the basename comes from the working directory.
-	let probe = if base_dir.as_os_str().is_empty() {
-		Path::new(".")
-	} else {
-		base_dir
-	};
-	let basename = probe
-		.canonicalize()
-		.unwrap_or_else(|_| probe.to_path_buf())
-		.file_name()
-		.map(|n| n.to_string_lossy().into_owned())
-		.unwrap_or_default();
-	sanitize_project_name(&basename)
-}
-
-/// Normalize a raw directory name into a valid compose project name: lowercase,
-/// keep only `[a-z0-9_-]`, then strip any leading `_`/`-`. Falls back to the
-/// `podup` literal when nothing valid remains, so the project name is never
-/// empty.
-fn sanitize_project_name(raw: &str) -> String {
-	let kept: String = raw
-		.to_lowercase()
-		.chars()
-		.filter(|c| c.is_ascii_lowercase() || c.is_ascii_digit() || *c == '_' || *c == '-')
-		.collect();
-	let trimmed = kept.trim_start_matches(['_', '-']);
-	if trimmed.is_empty() {
-		"podup".to_string()
-	} else {
-		trimmed.to_string()
-	}
 }
 
 #[tokio::main]
@@ -563,118 +234,4 @@ async fn run() -> podup::Result<()> {
 	}
 
 	Ok(())
-}
-
-#[cfg(test)]
-mod tests {
-	use super::{
-		resolve_base_dir, resolve_compose_files, resolve_project_name, sanitize_project_name,
-	};
-	use std::path::{Path, PathBuf};
-
-	#[test]
-	fn explicit_compose_files_win() {
-		let p = resolve_compose_files(&[PathBuf::from("custom.yml")]);
-		assert_eq!(p, vec![PathBuf::from("custom.yml")]);
-	}
-
-	#[test]
-	fn multiple_explicit_compose_files_preserved() {
-		let p = resolve_compose_files(&[PathBuf::from("a.yml"), PathBuf::from("b.yml")]);
-		assert_eq!(p, vec![PathBuf::from("a.yml"), PathBuf::from("b.yml")]);
-	}
-
-	#[test]
-	fn missing_compose_file_falls_back_to_default_name() {
-		// In a directory with no candidate files, the default name is returned
-		// so the resulting error names a sensible path.
-		let dir = std::env::temp_dir().join(format!("podup-cf-{}", std::process::id()));
-		std::fs::create_dir_all(&dir).unwrap();
-		let prev = std::env::current_dir().unwrap();
-		std::env::set_current_dir(&dir).unwrap();
-		// Guard against a COMPOSE_FILE set in the test environment.
-		let had_env = std::env::var_os("COMPOSE_FILE");
-		std::env::remove_var("COMPOSE_FILE");
-		let p = resolve_compose_files(&[]);
-		if let Some(v) = had_env {
-			std::env::set_var("COMPOSE_FILE", v);
-		}
-		std::env::set_current_dir(prev).unwrap();
-		let _ = std::fs::remove_dir_all(&dir);
-		assert_eq!(p, vec![PathBuf::from("docker-compose.yml")]);
-	}
-
-	#[test]
-	fn project_directory_override_wins() {
-		let base = resolve_base_dir(
-			Some(Path::new("/srv/app")),
-			Path::new("/etc/compose/docker-compose.yml"),
-		);
-		assert_eq!(base, PathBuf::from("/srv/app"));
-	}
-
-	#[test]
-	fn defaults_to_compose_file_parent() {
-		let base = resolve_base_dir(None, Path::new("/etc/compose/docker-compose.yml"));
-		assert_eq!(base, PathBuf::from("/etc/compose"));
-	}
-
-	#[test]
-	fn bare_filename_resolves_to_current_dir() {
-		let base = resolve_base_dir(None, Path::new("docker-compose.yml"));
-		assert_eq!(base, PathBuf::from(""));
-	}
-
-	#[test]
-	fn explicit_project_name_wins() {
-		let name = resolve_project_name(
-			Some("explicit".to_string()),
-			Some("from-compose"),
-			Path::new("/srv/myapp"),
-		);
-		assert_eq!(name, "explicit");
-	}
-
-	#[test]
-	fn compose_name_used_when_no_explicit() {
-		let name = resolve_project_name(None, Some("from-compose"), Path::new("/srv/myapp"));
-		assert_eq!(name, "from-compose");
-	}
-
-	#[cfg(unix)]
-	#[test]
-	fn falls_back_to_directory_basename() {
-		let dir = std::env::temp_dir().join(format!("podup-pn-{}", std::process::id()));
-		std::fs::create_dir_all(&dir).unwrap();
-		let name = resolve_project_name(None, None, &dir);
-		let _ = std::fs::remove_dir_all(&dir);
-		// The basename is sanitized; the temp dir name is already lowercase
-		// alphanumeric with hyphens, so it survives unchanged.
-		assert_eq!(
-			name,
-			dir.file_name().unwrap().to_string_lossy().to_lowercase()
-		);
-	}
-
-	#[test]
-	fn sanitize_lowercases_and_drops_invalid_chars() {
-		assert_eq!(sanitize_project_name("My App!"), "myapp");
-	}
-
-	#[test]
-	fn sanitize_keeps_underscore_and_hyphen() {
-		assert_eq!(sanitize_project_name("web_service-1"), "web_service-1");
-	}
-
-	#[test]
-	fn sanitize_strips_leading_separators() {
-		assert_eq!(sanitize_project_name("__leading"), "leading");
-		assert_eq!(sanitize_project_name("--dash"), "dash");
-	}
-
-	#[test]
-	fn sanitize_empty_result_falls_back_to_podup() {
-		assert_eq!(sanitize_project_name("!!!"), "podup");
-		assert_eq!(sanitize_project_name(""), "podup");
-	}
 }
